Made the public contacts CRUD private

+++ controllers/contact_controllers.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModels");
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id:req.user.id});
     res.status(200).json(contacts);
 });
 
@@ -12,7 +12,7 @@ const createContacts = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error("All fields are mandatory");
     }
-    const contact = await Contact.create({ name, email, phone });
+    const contact = await Contact.create({ name, email, phone,user_id:req.user.id });
     res.status(200).json(contact);
 });
 
@@ -31,6 +31,11 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact Not Found");
     }
+    if(contact.user_id.toString()!==req.user.id)
+    {
+        res.status(403);
+        throw new Error("User doesn't have permission to update somebody ele's contacts");
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -44,6 +49,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error("Contact Not Found");
+    }
+    if(contact.user_id.toString()!==req.user.id)
+    {
+        res.status(403);
+        throw new Error("User doesn't have permission to update somebody ele's contacts");
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
